restructure to suit adaptation
