+++ package.json
@@ -22,6 +22,7 @@
     "connect-flash": "^0.1.1",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
+    "dotenv": "^16.0.2",
     "ejs": "^3.1.8",
     "express": "~4.16.1",
     "express-session": "^1.17.3",
